[refs #81047] Fixing reindexing, indexer cleanup.
* Some functions appear as both indexers and object properties with the
same name. The catalog uses the properties for indexing and indexers do
not get called. Used best knowledge to merge differences between
indexer and methods with the same name, based on git history.

Applied a clear and rebuild to the catalog post changes and compared the
/statistics page for both pre and post recataloging, the data was
identical.

* Updated get_values_cat to use listFolderContents instead of
getFolderContents which is deprecated. This resolves the referenced
ticket.